[#488] Move std::hash implementation in separate file to fix clang-tidy scan bug
